fix small

+++ src/dynamics/attitude/ode_attitude_with_cantilever_vibration.hpp
@@ -26,16 +26,16 @@ class AttitudeWithCantileverVibrationOde : public InterfaceOde<13> {
     libra::Vector<13> output;
 
     libra::Vector<3> omega_b_rad_s;
-    for (int i = 0; i < 3; i++) {
+    for (size_t i = 0; i < 3; i++) {
       omega_b_rad_s[i] = state[i];
     }
     libra::Vector<3> omega_cantilever_rad_s;
-    for (int i = 0; i < 3; i++) {
+    for (size_t i = 0; i < 3; i++) {
       omega_cantilever_rad_s[i] = state[i + 3];
     }
 
     libra::Vector<3> euler_angle_cantilever_rad;
-    for (int i = 0; i < 3; i++) {
+    for (size_t i = 0; i < 3; i++) {
       euler_angle_cantilever_rad[i] = state[i + 10];
     }
 
@@ -49,26 +49,26 @@ class AttitudeWithCantileverVibrationOde : public InterfaceOde<13> {
 
     libra::Vector<3> rhs = inverse_inertia_tensor_ * (net_torque_b_Nm - inertia_tensor_cantilever_kgm2_ * angular_accelaration_cantilever_rad_s2);
 
-    for (int i = 0; i < 3; ++i) {
+    for (size_t i = 0; i < 3; ++i) {
       output[i] = rhs[i];
     }
 
-    for (int i = 0; i < 3; i++) {
+    for (size_t i = 0; i < 3; i++) {
       output[i + 3] = angular_accelaration_cantilever_rad_s2[i];
     }
 
     libra::Vector<4> quaternion_i2b;
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
       quaternion_i2b[i] = state[i + 6];
     }
 
     libra::Vector<4> d_quaternion = 0.5 * CalcAngularVelocityMatrix(omega_b_rad_s) * quaternion_i2b;
 
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
       output[i + 6] = d_quaternion[i];
     }
 
-    for (int i = 0; i < 3; i++) {
+    for (size_t i = 0; i < 3; i++) {
       output[i + 10] = omega_cantilever_rad_s[i];
     }
